Move all decoding code into text module
